feat: Add IP and geo information to resources
This commit adds the IP and geo information to the resources in the RegisterAPI class. The IP is obtained by mapping the axon IPs to the list of resources based on hotkeys. The geo information is retrieved using the IPWhois library. This enhancement improves the functionality of the RegisterAPI class.

Refactor the code to map axon IPs to resources and retrieve geo information using IPWhois library.